Add date to daily collection viewer
